Add OpenWiFi support for adrv9361z7035
A custom uEnv.txt file is automatically installed that loads in the system_top.bit file, which is provided by the user. Since OpenWiFi does not yet have .dtso files available, in case OpenWiFi package is selected, the dts directory and filename is changed to that of the OpenWiFi dts.